PB-170: Restructure code to show vision cone conditionally
